move exceptions into their own file
